equiv: tier-5 kfuncs, arena, may_goto, and the helper tail
Kfunc calls (relocations against undefined symbols) dispatch by name
onto the tier-2/3 patterns: rcu/preempt no-ops, release events,
acquire-style nullable object addresses (shared oracle scalars, so the
program's own NULL check is the fork), cpumask ops as event+oracle,
generic bpf_iter_<kind>_new/next/destroy, content-captured pure string
kfuncs, session cookie/is_return. bpf_throw is exact: it unwinds the
frame stack into the exception_callback: decl-tagged callback (or
returns the cookie), so rustc's decl-tag gap proves equivalent instead
of bailing. Arena: addr_space_cast is provenance-preserving identity,
alloc_pages forks into zeroed observable page regions, and pointer
stores into observable regions keep a shadow entry plus a canonical
identity encoding. may_goto is modeled never-taken (documented).
Helper tail: data-driven side-effect table (redirect/tunnel/sockopt/
storage_delete/...), buffer-writing env reads (get_stack, getsockopt,
tunnel key, check_mtu; errno clamped to the 0-or-negative contract),
copy_from_user with kernel-faithful zero-fill, seq_printf/seq_write,
sk_lookup/fullsock/tcp_sock. Model fixes from triage: per-run region
tags neutralized in pointer-identity captures; clang "func.var" and
rustc "var.N" local statics canonicalized to one region.

Corpus: 1020 EQUIV / 0 INEQUIV / 8 WAIVED / 6 UNKNOWN programs (254
objects fully proved; was 665/197 after CO-RE). Controls: mutated
string literal and seq_write length both flag INEQUIV via the trace.
True findings #7-#15 fixed and QEMU-gated: 0xabcd1234 unsigned wrap
(cgrp_ls_tp_btf); _Bool byte-compare class, clang emitting !=0 or ==1
per site (cgrp_ls x3, vrf_socket_lookup); struct-padding leak into a
map value (lsm_bdev); C pointer-arithmetic scaling kept faithfully
(test_sk_lookup_kern: tuple+sizeof*tuple=+1296, sk+=1=+80); 4-byte
store through __u64* (test_skmsg_load_helpers); unsigned-promotion
compare (test_sockmap_strp); copy_from_user result discarded
(test_spin_lock_fail); int sign-extension into a u64 helper arg
(xdp_redirect_multi_kern); 62 dropped bpf_printk log sites restored
(test_tunnel_kern, test_sk_lookup).

+++ progs/test_skmsg_load_helpers.rs
@@ -76,16 +76,18 @@ fn prog_msg_verdict_common(msg: *const sk_msg_md) -> i32 {
     let pid = (bpf_get_current_pid_tgid() >> 32) as u32;
     let sk = unsafe { (*msg).sk } as *const c_void;
 
+    // C: sk_stg is __u64* — the store writes all 8 value bytes (pid
+    // zero-extended), not 4; a u32 store would leave residue in the map.
     let sk_stg = bpf_sk_storage_get(
         &socket_storage,
         sk,
         core::ptr::null(),
         BPF_SK_STORAGE_GET_F_CREATE,
-    ) as *mut u32;
+    ) as *mut u64;
     if sk_stg.is_null() {
         return SK_DROP;
     }
-    unsafe { *sk_stg = pid };
+    unsafe { *sk_stg = pid as u64 };
 
     let mut tpid: u32 = 0;
     let tgid_addr = unsafe { &*task }.tgid().as_ptr();
